Bug/VETS-CPC-842_Fix_Refresh_of_Add_Review_to_Vet (#422)
JIRA: [link to jira
ticket](https://champlainsaintlambert.atlassian.net/jira/software/c/projects/CPC/boards/13?selectedIssue=CPC-842)
## Context:
Before, when adding the rating for a vet, you had to manually refresh
the page to make it appear.
## Changes
Fixed refresh when you add a rating for a vet. Now it automatically
refreshes when you click the submit button. Additionally, I fixed the
String for the date when you submit a description to a readable format.
## Before and After UI (Required for UI-impacting PRs)

## Dev notes (Optional)
- Specific technical changes that should be noted
## Linked pull requests (Optional)
- pull request link