Bug/bill cpc 746 change display id to names (#451)
JIRA: link to jira ticket
## Context:
The main page for bills (bill history) displayed a user Id and vet Id
which meant nothing (they aren't the ones used as parameters in
requests). So I switched them to instead display the owner's and vet's
full names. Also had to change the type of these two attributes to match
the type of their corresponding attribute in their respective
microservice. This was done to facilitate later changes involving the
two attributes.
## Changes
- Owner Name column added
- Owner Name column now displays owner's full name 
- Vet ID column added
- Vet Name column now displays vet's full name 
- CustomerId is now a string instead of an integer throughout the
billling service and api-gateway
- VetId is like customerId now a string instead of an integer.
- Added a method to get the vet entity through matching the vetBillId
attribute of the vet class and the vetId attribute of the bill class
- Added a method to get the customer's name through mapping the names
with an id and then comparing that id to the customerId bill holds
## Before and After UI (Required for UI-impacting PRs)
Before
![Screenshot 2023-09-24
201841](https://github.com/cgerard321/champlain_petclinic/assets/43860533/faaf7a96-fc61-4149-b7bf-4e98bbbf619e)
After
![Screenshot 2023-09-25
193929](https://github.com/cgerard321/champlain_petclinic/assets/43860533/b85343eb-23b7-40ef-a791-0c58d20d2a17)