feat(VIST-CPC-748): full story exception handling fix front end (#469)
JIRA: https://champlainsaintlambert.atlassian.net/browse/CPC-748
## Context:
Visits-service had no exception-handling whatsoever. Not checking what
is being sent is awful since the front end can basically send in
anything and our service will still accept it.

Example: 

Here is the request that was being sent by the front end:

![image](https://github.com/cgerard321/champlain_petclinic/assets/92550620/780c6304-d84b-428c-8f17-1b60c8d77d2c)

Now here is our requestDTO:


![image](https://github.com/cgerard321/champlain_petclinic/assets/92550620/b4dc70ef-cccf-47b4-a672-6920438b1c24)

## Changes
What are the various changes and what other modules do those changes
effect.
This can be bullet point or sentence format.


By implementing the exception handling, I was able to understand that
the visitDate is null since what is being passed in the request is
"date" and not "visitDate". It is also badly formatted, so it wouldn't
work even if it was called visitDate
Also, there is no petId in this post request. The post was working but
it was wrong.

Test Coverage for Visits-Service:

![image](https://github.com/cgerard321/champlain_petclinic/assets/92550620/c1ab126c-c681-4eb3-8094-3a30635a3047)

The main key to making this exception-handling work was to get the
api-gateway to actually hold the exception that was thrown by our
service and pass it on. In order to get the message from visit-service
to the front-end, I parsed the error return of the visits-service to the
api-gateway on reception:


![image](https://github.com/cgerard321/champlain_petclinic/assets/92550620/2f01936a-5340-4eec-92f9-354e1e5fdcaf)


## Before and After UI (Required for UI-impacting PRs)
Now, instead of only displaying a green notification for the successful
creation of a visit, it will also display a red notification and tell
the user exactly what is wrong with the completion of the form.

**BEFORE UI**

![image](https://github.com/cgerard321/champlain_petclinic/assets/92550620/7ada31e4-3892-4514-9fcc-6e66aac6c7fb)

There was also a required annotation on the input fields in case the
user didn't insert anything in them:

![image](https://github.com/cgerard321/champlain_petclinic/assets/92550620/c3bbc7fe-3112-4b7c-9d3a-a0fac635cfd3)


**AFTER UI**
Example for visit without description:


![image](https://github.com/cgerard321/champlain_petclinic/assets/92550620/da080c9e-6549-4c07-8f7b-79b4f134e9b0)

Example for null visitDate:

![image](https://github.com/cgerard321/champlain_petclinic/assets/92550620/95129f19-4b90-4367-bc59-483d328706d1)

Now even if complete everything in the form, although the date has been
chosen properly here, as explained earlier, the request is being sent in
wrong, therefore visit date is null:


![image](https://github.com/cgerard321/champlain_petclinic/assets/92550620/2c16f830-59ed-4d44-af84-416ccd228473)

I am unable to display what would happen past this since the request has
to be fixed if i want to show all the exceptions. But here they are in
my service:

![image](https://github.com/cgerard321/champlain_petclinic/assets/92550620/124e0270-7983-4378-84bf-cd3b109ca46f)