feat(VETS-CPC-837): add pre-defined review for a rating to make user feedback more flexible and friendly (#499)
JIRA: link to jira ticket
CPC-837

https://champlainsaintlambert.atlassian.net/jira/software/c/projects/CPC/boards/13?selectedIssue=CPC-837&sprints=313
## Context:
This ticket is about including a pre-defined review for when a user add
or update a rating of a specific vet. Previously, the user was only able
to write a description (explanation) about the rate score they gave for
the vet. Now, they will be able to select it as well, if they want to.
This makes it easier for the users who want to give written feedback but
do not/cannot type it themselves.
## Changes
- Change made in Rating entity, Rating Response DTO, Rating Request DTO
to include "private PredefinedDescription predefinedDescription"
- Change made in addRatingToVet method in the vets-service to include
the possibility of a selected pre-defined review description.
- Change made in updateRatingByVetIdAndRatingId method in the
vets-service to include the possibility of a selected pre-defined review
description.
- Change made in vet-details.controller.js and vet-details.template.html
to include the predefinedDescription and its display.
- Added tests in VetsServiceClientIntegrationTest and
ApiGatewayControllerIntegration test, as well as in RatingRepositoryTest
and VetControllerIntegrationTest
- Changed methods (add rating and update rating) are tested > 90%.
## Before and After UI (Required for UI-impacting PRs)
Before

![image](https://github.com/cgerard321/champlain_petclinic/assets/46633364/a65cf6d4-9e99-4e24-9af6-6978e580ecf3)
After

![image](https://github.com/cgerard321/champlain_petclinic/assets/46633364/77cca910-e786-43aa-bd2f-b270eafe959e)

![image](https://github.com/cgerard321/champlain_petclinic/assets/46633364/314ad50c-80c8-4bd7-9690-f01cc1ef2b05)
## Dev notes (Optional)
- Specific technical changes that should be noted
## Linked pull requests (Optional)
- pull request link