Feat/INVT-CPC-889 search letter or number result (#526)
JIRA: link to jira ticket:
https://champlainsaintlambert.atlassian.net/browse/CPC-899
Added the ability to search a product by a letter or a number
-After further review with my teammate I have removed all the
unnecessary code unused.
-Modified the back-end for Inventory and product get all to regex which
enables the inventory to appear depending on the letter and number
inputed.
-Added a button that will facilitate the search better which is the
Enter key button rather than just clicking the search

Before (By the entire name needed to be written):

![image](https://github.com/cgerard321/champlain_petclinic/assets/91092017/4e9ec52c-53ac-45e0-a1a4-281b3bb89289)

Before searching by letter or number for inventory:

![image](https://github.com/cgerard321/champlain_petclinic/assets/91092017/dc97e75b-9d7a-4461-85a9-10602d898f2f)
After putting a letter:

![image](https://github.com/cgerard321/champlain_petclinic/assets/91092017/b02cd914-df4d-49f2-8aa0-a38650ed1f6c)
After inserting a number:

![image](https://github.com/cgerard321/champlain_petclinic/assets/91092017/fcf1f7b7-463f-4731-9dbd-b79d40002123)
For description:

![image](https://github.com/cgerard321/champlain_petclinic/assets/91092017/893d1fd5-df92-41d4-b0f7-46444a99d718)

Before Product:

![image](https://github.com/cgerard321/champlain_petclinic/assets/91092017/ea49e6ef-8523-45a2-84db-66d274b1fc2e)

After inserting the product letter or numeric digit for the name:

![image](https://github.com/cgerard321/champlain_petclinic/assets/91092017/62be0bbc-2730-46d8-8142-439a88a4a61f)

![image](https://github.com/cgerard321/champlain_petclinic/assets/91092017/2649a22d-6a9c-4218-b189-4619c4febcdc)

-Did not implement this feature for the product price and product
quantity because regex does not support double or Int. If I were to
replace all double or int to string it would affect and create a bunch
of merge conflicts.I would also need to change all the database double
and int type to string too.
- pull request link

Took 2 minutes

Took 7 minutes