Feat/invt cpc 981 access product list through inventory row (#532)
JIRA: link to jira ticket
https://champlainsaintlambert.atlassian.net/browse/CPC-989       
                   
What is the ticket about and why are we making this change?

In this pull request, I made tweaks to the front end to enrich the user
experience. Instead of clicking on the name of the inventory to view the
product inside of it now you can simply click on the row itself with the
hover on the mouse feature that I added as well. We are making this
change because it makes the flow of the inventory more user-friendly and
easier to navigate through the inventory.

What are the various changes and what other modules do those changes
affect?

-Clicking on the row of an inventory links you to the product list
inside of that inventory
-CSS: added hover on the mouse that appears on the inventory rows
-Made the edit and delete buttons inside the rows work properly with the
new feature I added

## Before and After UI (Required for UI-impacting PRs)

Before: 

![image](https://github.com/cgerard321/champlain_petclinic/assets/57909059/e4a07ba0-9b65-406a-9f62-322629903f5e)

After: Mouse Hover feature

![image](https://github.com/cgerard321/champlain_petclinic/assets/57909059/46a980c4-7121-4cd4-bd92-3b03afd13788)

After: clicking anywhere on the selected row brings you to the Product
List

![image](https://github.com/cgerard321/champlain_petclinic/assets/57909059/3493084b-6456-45f6-8dec-dd214e3a661e)


## Dev notes (Optional)
- We go Gym