Feat/invt cpc 979 add sale price for clients in product list (#533)
JIRA: link to jira ticket
https://champlainsaintlambert.atlassian.net/browse/CPC-979

What is the ticket about and why are we making this change?

This ticket is about adding an important field to our product list
called the sale price. The sale price is important because previously we
only had the normal price for a product that the pet clinic bought it
for and we were missing the price we will be selling it at. We are
making those changes because it's crucial information about a product.

What are the various changes and what other modules do those changes
affect?

- The front end has been changed to accommodate the new field in the
product list.
- A couple of things about the backend have been changed: 
1. RequestDTO 
2. ResponseDTO
3. Product Entity
4. ServiceImpl methods

- All the Testing that was using the DTOs, the Product entity and the
serviceImpl has been changed to accommodate the new changes.
- API Gateway testing fully accommodated with the new changes.

## Before and After UI (Required for UI-impacting PRs)
Before: 

![image](https://github.com/cgerard321/champlain_petclinic/assets/57909059/943b10f7-9cb9-47ad-9fec-a88bc58bf2ab)

After:

![image](https://github.com/cgerard321/champlain_petclinic/assets/57909059/4e42e629-f35b-4b46-8834-6a32636929d2)