Feat/invt cpc 993 ability to search product by sale price (#548)
JIRA: link to jira ticket
https://champlainsaintlambert.atlassian.net/browse/CPC-993

**What is the ticket about and why are we making this change?**

This ticket is about adding the ability to search for a product or
multiple products by the new field Sale Price. We are making this change
because it's important to be able to search by sale price since we are
already searching by regular price it's the logical thing to do to allow
that search. Searching by sale price is important because there could be
a lot of different sale prices and sorting by that field would make it
easier to filter out the products more precisely.

**What are the various changes and what other modules do those changes
affect?**

- Frontend fully working and integrated
- Backend fully working and integrated into the service and API-Gateway
- Tests that were already there were changed in order to accommodate the
new needed type of search
- All 4 types of testing were added for the search:
1. ServiceUnitTest
2. RepositoryTest
3. ControllerIntegrationTest
4. ControllerUnitTest

## Before and After UI (Required for UI-impacting PRs)
Before: 

![image](https://github.com/cgerard321/champlain_petclinic/assets/57909059/49539f8a-2191-4b9a-94dd-9d4007c73696)

After:

![image](https://github.com/cgerard321/champlain_petclinic/assets/57909059/ce20d504-105f-4f48-b448-465994717b84)

When searching for inventoryId and Sale Price alone:

![image](https://github.com/cgerard321/champlain_petclinic/assets/57909059/b7f6d03d-c87a-48df-b17f-efff13423ceb)

When searching with all the fields including sale price:

![image](https://github.com/cgerard321/champlain_petclinic/assets/57909059/9f20c9e3-0217-4b9f-907f-7ed5aeb27886)

When searching by 3 fields including the sale price:

![image](https://github.com/cgerard321/champlain_petclinic/assets/57909059/5edcc247-a5f6-480f-aa38-dddcc99f74cc)