feat(VIST-CPC-927): implementing aggregation and details page (#579)
JIRA:[ link to jira
ticket](https://champlainsaintlambert.atlassian.net/browse/CPC-927)
## Context:
Visits only had access to the petId and vetId. By aggregating it, we now
have access to the needed vet information and pet information. There was
also no details page for our visits. The user only had access to the
general view which is a list.
This also allows us to populate a details page with a single get method.
## Changes

-Changed visitResponseDTO to now have the vet's full name and contact
information as well as the pet's birth date and name.
-Added navigation to Details Page
-Added Details Page to the front-end

UI CHANGES:

**BEFORE:**


![image](https://github.com/cgerard321/champlain_petclinic/assets/92550620/b21ea8dd-876d-41b1-ae55-c9509b01ffde)

**AFTER:**


![image](https://github.com/cgerard321/champlain_petclinic/assets/92550620/d61f8f3a-fe15-480e-aa19-c2faf76429de)

DETAILS PAGE:

![image](https://github.com/cgerard321/champlain_petclinic/assets/92550620/5c98e72e-54ad-456a-a1d9-0791c04520ad)