Create parallel_microservice_build.yml (#606)
JIRA: None
## Context:
Seperated the builds of the microservices across different to improve
build times
## Changes
- Ran the builds invividually in github actions
- Did not remove the last build job to try and see if the seperation
improves the build speed
## Dev notes (Optional)
- It seems the qodana is getting faster and faster each build because of
the caching so i can't see if the change actually improves speed. I
suggest we keep these actions optional. If after students try and use it
and we see that it is slower then the qodana. I can try and seperate it
in qodana. Maybe the runners they are using might have more resource and
then the builds can be quicker. Implemented it in the github actions to
keep it all in one place and you don't need an external account for the
github actions so students can go see the config and change it
themselves.
- Added gradlew.bat file to the auth-service because it didn't have one
like the other microservices