Integration auth new fe (#607)
JIRA: none :(
## Context:
Addition of axios error handling logic.
Addition of placeholders for students to add the required error pages
for when requests return bad request, not found, bad gateway....
## Changes
- Updated the code to integrate the basic authentication page to be able
to authenticate the user on a basic HTML page.
- Added the url of the nginx from the hosted frontend to not get cors
error
## Dev notes (Optional)
- Students need to make error pages for the different response statuces
that the api will return these need to be made in the page folder.
- Students need to make a the forgot password page.
- Students need to also add all the default pages.

---------

Co-authored-by: dylanbrass <dylan.brassard@outlook.com>