Added the Products Service into the parrallel workflow, added at the … (#617)
…bottom of the file

JIRA: https://champlainsaintlambert.atlassian.net/browse/CPC-1109
## Context:
This ticket was about fixing a configuration where we forgot to add the
new service we created (products-service) into the parallel workflow
inside the .github folder.

## Changes

Towards the bottom of the file I created a new profile called
products-service-test. This has the same configurations as all the other
tests produced in that section.