Feat/visit cpc 1112 add a new exception in case the email is already sent (#620)
[JIRA: link to jira
ticket](https://champlainsaintlambert.atlassian.net/browse/CPC-1112)
## Context:
We are adding a new exception before creating the mailing system so that
both service can communicate more easily. Did not implement the
exception yet. Did is just in preparation for future changes
## Changes
We first added a new exception. It should not affect any services as it
was not implemented and is only related to visits.
Also added some documentation on some functions for the sake of clarity
as a team. ( Only one person was on the branch )