feat(VETS-CPC-1110): make veterinarian and home page accessible. (#619)
JIRA: link to jira ticket
## Context:
What is the ticket about and why are we doing this change.
## Changes
What are the various changes and what other modules do those changes
effect.
This can be bullet point or sentence format.
## Before and After UI (Required for UI-impacting PRs)
If this is a change to the UI, include before and after screenshots to
show the differences.
If this is a new UI feature, include screenshots to show reviewers what
it looks like.
## Dev notes (Optional) 
##### (The next part is an addition from @nic5694 to help with the
issues that some people have been running into in the past couple of
days)
Added workspace configurations to VS Code because some people were
missing plugins needed to see linting errors they might introduce. This
update includes settings for format on save—when you modify a file and
save it, VS Code will automatically format the code for you.

**Important**: If you already have a .vscode folder in the root of the
petclinic-frontend folder, remove it before pulling from main, or you’ll
get a conflict. It’s best that everyone uses this configuration to
ensure smoother development and avoid build failures in CI.

**Extension Recommendations**:
Added recommended extensions that you’ll find at the bottom of the
Extensions Marketplace in VS Code. Please **install the following**:
        Prettier: For consistent code formatting.
        ESLint: To catch linting errors.
"try-material-icon-theme": To make it easier to navigate the folder
structure with icons.

These tools will help you spot issues early and streamline the
development process.
## Linked pull requests (Optional)
- pull request link

---------

Co-authored-by: nic5694 <nicholasmartoccia04@icloud.com>