Created a review entity in Visit, created the service and service imp… (#653)
…l. Done the test. In the api gateway, addeded the review. In the front
end, add, update and get all reviews

https://champlainsaintlambert.atlassian.net/browse/CPC-1117
It's about creating a review section for the users, after a visit, they
can write what they thought on the visit.
If the PR changes the .vscode folder, explain why in detail because it
should not.
Be sure to include cgerard321 as a reviewer.

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

What are the various changes and what other modules do those changes
affect.
This can be bullet point or sentence format.
- New Review entity in the Visit-service
- new service and service impl for review
- New testing for the crud methods
- New method in the visitserviceClient in the api-gateway
- New methods in the visitController in the api-gateway
- Front-end for add, update and display the reviews
- new navigations for reviews

If this is a change to the UI, include before and after screenshots to
show the differences.
If this is a new UI feature, include screenshots to show reviewers what
it looks like.

![Screenshot 2024-09-17
190321](https://github.com/user-attachments/assets/400bcb79-b079-4ab7-8b49-1ddb915602ee)

![image](https://github.com/user-attachments/assets/faa2da54-1f1c-4eb6-949d-8847531dc3e8)

Specific technical changes that should be noted
Pull request links

Co-authored-by: Zakaria_ <135060272+tramway23@users.noreply.github.com>