Feat/vist cpc 1117 add reviews section in admin page (#665)
**JIRA:** link to jira ticket
https://champlainsaintlambert.atlassian.net/browse/CPC-1117
It's about creating a review section for the users, after a visit, they
can write what they thought on the visit.
If the PR changes the .vscode folder, explain why in detail because it
should not.
Be sure to include cgerard321 as a reviewer.

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

What are the various changes and what other modules do those changes
affect.
This can be bullet point or sentence format.

- Did the test in Api-gateway. for the vistiController and
visitServiceClient
-  Did the css for the forms
- Fixed some issues and mistakes in the front-end

If this is a change to the UI, include before and after screenshots to
show the differences.
If this is a new UI feature, include screenshots to show reviewers what
it looks like.

Before:

![image](https://github.com/user-attachments/assets/3887df99-d2d9-43a3-a218-68bd0dec717b)

After:

![image](https://github.com/user-attachments/assets/0a60960b-fcbd-40a0-b03a-c3349d8e706c)

Specific technical changes that should be noted
Pull request links

---------

Co-authored-by: Zakaria_ <135060272+tramway23@users.noreply.github.com>