Feat/vets cpc 1076 get vet by vet name (#684)
**JIRA:** link to jira ticket
https://champlainsaintlambert.atlassian.net/browse/CPC-1076
## Context: added the full functionallity to find a vet by their first
or last name using a searchbar
What is the ticket about and why are we doing this change.
we fully implemented a searchbar in our frontend and backend to allow
you to search a vet by their first and last name
## Does this PR change the .vscode folder in petclinic-frontend?: no
If the PR changes the .vscode folder, explain why in detail because it
should not.
Be sure to include cgerard321 as a reviewer.

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
What are the various changes and what other modules do those changes
affect.
This can be bullet point or sentence format.
- we now have a fully functionning searchbar that allows you to search
for a vet by first or last name
## Before and After UI (Required for UI-impacting PRs)
If this is a change to the UI, include before and after screenshots to
show the differences.
If this is a new UI feature, include screenshots to show reviewers what
it looks like.

Before:

![image](https://github.com/user-attachments/assets/3ac35a8d-7585-488d-a7e3-2b505e2c2966)

After:

![image](https://github.com/user-attachments/assets/ce0a299f-7cf4-49f0-8f88-bca8aa7631f7)

## Dev notes (Optional)
Specific technical changes that should be noted
## Linked pull requests (Optional)
Pull request links