feat/(VIST-CPC-1125): implement add visits (#706)
**JIRA:** [link to jira
ticket](https://champlainsaintlambert.atlassian.net/browse/CPC-1125)
## Context:
What is the ticket about and why are we doing this change.
Implementation of Add Visit to the Visits section. We are doing this
change because previously it wasn't available to create a Visit.
This allows to create a visit in the new endpoint created (/visits/add)
## Does this PR change the .vscode folder in petclinic-frontend?:
It doesn't make changes to .vscode folder.

## Changes
What are the various changes and what other modules do those changes
affect.

Added a new POST controller to add a visit in /api/v2/gateway/visits
Implemented logic in the VisitsServiceClient to create a new visit.
Wrote unit tests to make sure the creation of the new visit is well
done.
Added a button to go to the add visit form.

## Before and After UI (Required for UI-impacting PRs)
Before:

![{07992AD4-FB1D-4FBA-AC59-E7C2B6278451}](https://github.com/user-attachments/assets/9dabecd3-7983-4b52-94e0-79781e76a5d7)
After:

![{2610936A-4009-4AB1-95DD-6379B7EFFC61}](https://github.com/user-attachments/assets/3518e7d5-82b7-44b9-82dd-3b8dd9ee6778)

## Dev notes (Optional)
In order to get the right time, I had to subtract 4 hours from it
because for some reason it was adding 4 hours to the selected time. So I
subtract it from the request (basically for the back-end) and the
front-end, so it displays the right time.