feat(VETS-CPC-1091) add photos by vet (#728)
**JIRA:** link to jira ticket
## Context:https://champlainsaintlambert.atlassian.net/browse/CPC-1091
What is the ticket about and why are we doing this change.

## Does this PR change the .vscode folder in petclinic-frontend?: no
If the PR changes the .vscode folder, explain why in detail because it
should not.
Be sure to include cgerard321 as a reviewer.

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
What are the various changes and what other modules do those changes
affect.
This can be bullet point or sentence format.
-changed the photoService,photoServiceImpl, vetsController, apigatewayv2
controller and the frontend mainpage for vets
## Before and After UI (Required for UI-impacting PRs)
Before:

![oldui](https://github.com/user-attachments/assets/2d49c6cb-ba63-4d84-a968-95d466d03ba1)
After:

![successfullUpload](https://github.com/user-attachments/assets/2ba0f0b4-d97e-42fc-9abb-2aed80cab944)

If this is a change to the UI, include before and after screenshots to
show the differences.
If this is a new UI feature, include screenshots to show reviewers what
it looks like.
## Dev notes (Optional)
Specific technical changes that should be noted
-i changed a couple of methods in the controller related to photos but
it should'nt affect any member of my team

## Linked pull requests (Optional)
Pull request links