Feat/invt cpc 1050 adding specific supplies to inventories (#748)
**JIRA:**
https://champlainsaintlambert.atlassian.net/jira/software/c/projects/CPC/boards/15/backlog?selectedIssue=CPC-1050
## Context:

-This ticket is about implementing an endpoint to add specific supplies
to inventories.
-This ticket also introduces a supply entity which we intend to use
instead of the Product entity.

## Does this PR change the .vscode folder in petclinic-frontend?:
No
## Changes
One of the big changes is the addition of a supply entity which our team
intends to use in the future in place of the product entity.
Additionally, three endpoints were added.

•One for adding a supply to an inventory.
•One for getting all supplies from an inventory by it's inventory name.
•One for getting all inventory names.
## Before and After UI (Required for UI-impacting PRs)
Before we only had a button called Add Supply which did not do anything
when you clicked on it:

![image](https://github.com/user-attachments/assets/2bac8efc-c5d4-443e-91ce-e6feababfb3f)

After we have a fully functional Add Supply Form which let’s us add a
supply by inventory name. And the inventory names aren't hard coded they
are fetched from our database.

![image](https://github.com/user-attachments/assets/14a44b62-54d5-47bf-b369-e7825d9ac8f1)

![image](https://github.com/user-attachments/assets/b55335e8-e4d9-4644-80d2-668045fe1d1a)