Feat/INVT-CPC-1040/Display_Low_Stock_Supplies_On_Click (#754)
[**JIRA:** link to jira
ticket](https://champlainsaintlambert.atlassian.net/browse/CPC-1040)
What is the ticket about and why are we doing this change.
The ticket is about adding a button that displays every supplies that is
considered to have low stocks. It displays the supplies that will need
to be ordered.

What are the various changes and what other modules do those changes
affect.
It adds a check low stock for all inventory button and when you click on
it a list of supplies that are low stocks are displayed. Those changes
don't affect any other modules only the inventory.


![CPCChangesLowStock](https://github.com/user-attachments/assets/15ccf7cf-0502-4d2d-9b2a-bf8b78f6b457)