Feat/prod cpc 1030 add search by product name front end3001 (#763)
**JIRA:** : https://champlainsaintlambert.atlassian.net/browse/CPC-1030

## Context:
This feature allows people to search for a specific product by its name.
It is not case sensitive. When clikcing on the name of the product that
pops up it will display all of the products information.

This feature does not enteract with the .vscode folder in the front-end.


## Changes

Added another component and css file that should search for a specific
product. I then created an instance of the component inside our products
page.

## Before and After UI (Required for UI-impacting PRs)

![Screenshot 2024-09-23
112643](https://github.com/user-attachments/assets/71786148-ee56-43ee-96a8-3dc5a02531bf)