bug(VETS-CPC-1249): Fixed bug with update vet. (#785)
**JIRA:** https://champlainsaintlambert.atlassian.net/browse/CPC-1249

## Context:
The page didn't reload automatically when doing an update for a vet so
we had to manually reload the page to see it. This is bad practice so I
added a autotmatic relaod upon saving a vets upated info. I also
realised that the "Specialities" field was missing in the form so I
added it.

## Does this PR change the .vscode folder in pet clinic-frontend?:

This PR does not effect the .vscode file.


## Changes

- Added a automatic page relaod after updating vet.
- Added the missing "Specialities" field to the form.

## Before and After UI (Required for UI-impacting PRs)
for the missing field in the form : 
Before: 
The field did not appear.

After:

![image](https://github.com/user-attachments/assets/ede75fe4-81c0-4505-9235-f6d70d10657a)


## Dev notes (Optional)
Specific technical changes that should be noted
## Linked pull requests (Optional)
Pull request links