bug(VETS-CPC-1248): Fix delete vet by vet ID (#790)
JIRA:
https://champlainsaintlambert.atlassian.net/browse/CPC-1248?atlOrigin=eyJpIjoiYmRmMDM1NmMyNjE4NGY1ZmE0OGJlMDQxZWQ5NzA2OWYiLCJwIjoiaiJ9

## Context:
What is the ticket about and why are we doing this change.


This ticket is about deleting a veterinarian by their ID, the admin
needs to have this ability to ensure that the list of vets is
up-to-date, especially when a vet can quit or be fired.

So this change fixes a problem where the vet was not properly deleted
from the database. Also, it fixes a bug where after clicking the
"delete" button, it was navigating the admin to the vet's detail page
instead of staying on the list.


## Does this PR change the .vscode folder in petclinic-frontend?:

No, this PR does not modify the .vscode folder

## Changes

- Fixed the delete functionality so that it properly deletes the
veterinarian record from database (MongoDB).

- Addressed a bug where, after deleting a vet, the UI would navigate to
the vet's detail page. Now, the page stays on the veterinarian list and
removes the deleted vet from the UI.

## Before and After UI (Required for UI-impacting PRs)

BEFORE:

1- Initial UI with "John Doe":


![image](https://github.com/user-attachments/assets/03f1916d-cdaf-4eaa-9f7c-d2b8f2ed2d62)

2- Attempting to Delete "John Doe": 


![image](https://github.com/user-attachments/assets/abbdd5af-bb8d-459f-ad0f-209f3c366981)

3- Reloading the page still shows the deleted vet: 


![image](https://github.com/user-attachments/assets/796b7c28-6cd5-44ee-83f7-80aef24c5164)


AFTER:

1- Initial UI with "John Doe": 


![image](https://github.com/user-attachments/assets/c6e19f60-db2d-4d63-a5b0-4765110ae6a3)

2- Approval to delete this vet:


![image](https://github.com/user-attachments/assets/35c7c55a-06cd-48b9-8e8d-13f0b69b620d)

3- Deleted "John Doe" from the UI:


![image](https://github.com/user-attachments/assets/85887c47-0e15-4f47-afa0-41c0282df5f9)

4- MongoDB reflecting the deletion:


![image](https://github.com/user-attachments/assets/b6d0bec1-3692-49bf-ab70-28ad08385f5d)



## Dev notes (Optional)
- this fix ensures the deletion in the backend and after deletion the
vet list is updated in real-time without refreshing the page