bug(VETS-CPC-1276): bug fix for add vet (#791)
**JIRA:** https://champlainsaintlambert.atlassian.net/browse/CPC-1276

## Context:
After adding a vet, we had to manually reload the page to be able to see
the newly added vet. So I added an automatic reload for the page once a
new vet was added. I also realised that the form for the add did not
match the one for the update. So I decided to change them both to set
all fields in a logical order.

## Does this PR change the .vscode folder in petclinic-frontend?:
This PR does not effect the .vscode file.

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes

- Added an automatic reload to the page once a new vet was added.
- Re-organized the form for add and update so that they are the same and
fields are in a logical order.

## Before and After UI (Required for UI-impacting PRs)
Before: 

- Page didn't reload
- Fields in the forms where not in the same order. 

After:

Field in add and update form are in the same order:

![image](https://github.com/user-attachments/assets/4c95f009-0175-4143-8291-d7ca35f9aa4d)

![image](https://github.com/user-attachments/assets/896192ed-80ba-4db9-a91e-1806885513af)

Pages relaods on its won after adding a new vet.

## Dev notes (Optional)
Specific technical changes that should be noted
## Linked pull requests (Optional)
Pull request links