Feat(CUST-CPC-1188): Users can register on the website using a sign up page (#796)
[**JIRA:** link to jira
ticket](https://champlainsaintlambert.atlassian.net/jira/software/c/projects/CPC/boards/15/backlog?selectedIssue=CPC-1188)
## Context:
What is the ticket about and why are we doing this change.
## Does this PR change the .vscode folder in petclinic-frontend?:
It doesn't do any changes to the vscode folder

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
- Added sign up page and relevant css
- Created V2 functions in the api gateway for createUser
- Created V2 functions in the api gateway for verifyUser
- Added EmailSendingFailedException
- Added 3 tests in new class v2/AuthControllerUnitTest
- Added 3 tests in new class
v2/domainclientlayer/AuthServiceClientControllerIntegrationTest
- Added 1 test in new class businesslayer/AuthServiceUserServiceTests
- Changed from using the old emailing service to the new one to send
verification email
- Removed a cart post api request that was added in the UserServiceImpl
in auth service as the person implementing added it and pulled without
testing and it never functioned, so it was breaking sign up. I have
contacted and will be fixing this problem in a future pull request.
- Added Models in the backend to communicate with the new emailing
service (DirectEmailModelRequestDTO and EmailModelResponseDTO) as well
as added the domain client for it
- Added Models needed in the front end to send the user and customer
details to the backend (RegisterModel and OwnerRequestModel)
- Changed the redirection of the verification link to the new front end
after the token is verified
- Changed the relevant links on the website to go to the sign up page
- Made the page reload on logout so it deletes user info and you can't
just press back and be logged in again
- temporarily reduced jacoco testing minimum to 85 in auth service
## Before and After UI (Required for UI-impacting PRs)
Before there was no sign up page, now:

![image](https://github.com/user-attachments/assets/bdb8a350-9eb6-4c7e-b145-29d9a1ff734c)

![image](https://github.com/user-attachments/assets/9194626c-1288-4ee3-997d-4e9708debf1b)

It also has instant validation, if you modify a field like write
something and remove it, it will tell you a field is required, and as
you write your password it will tell you what is required. Max limit of
character for everything but email: 60, max for email: 320, Phone needs
to be an uninterrupted 10 numbers, email is checking for the email
format and password is checking for "Password must contain at least 8
characters, including an uppercase letter, a lowercase letter, a number,
and a special character."

![image](https://github.com/user-attachments/assets/af5d7a90-5147-4116-96cf-513723beeea6)

Loading:

![image](https://github.com/user-attachments/assets/9ef63a78-fb2f-48f1-8353-f218d059abef)


Here is a successful email being sent:
On the user's end:

![image](https://github.com/user-attachments/assets/f1de76ef-b261-4292-8664-2809a869532a)

On the emailing end:

![image](https://github.com/user-attachments/assets/1f7d8174-2183-48ea-848d-2ec5f5de2682)


## Dev notes
<span style="color:red">**IF YOU ARE USING AN ANTIVIRUS**</span>
The antivirus might be blocking your access to the mailing API, if the
email is not being sent after submitting the form, please make sure to
either grant access to the api in your antivirus (The SMTP server) or
disable it while you are trying to sign up.

The auth-service minimum jacoco testing has been temporarily lowered to
85, I am waiting for another PR to be able to do the remaining tests