feat(INVT-CPC-1271_Add_Supply_Inside_Inventory) (#815)
**JIRA:https://champlainsaintlambert.atlassian.net/browse/CPC-1271
## Context:
What is the ticket about and why are we doing this change.
-Implemented the logic for adding a suplly in the back end 
-The front end displays a button called add that allows the admin to add
a supply inside an inventory
## Does this PR change the .vscode folder in petclinic-frontend?:No
## Changes
What are the various changes and what other modules do those changes
affect.
-Added a add button 
-The add button will display a form, when the add will submit the form,
the new suplly will be added to the inventory
## Before and After UI (Required for UI-impacting PRs)
Before

![image](https://github.com/user-attachments/assets/97c9b6a8-7a33-4075-9a5c-3ce517016a2e)
After

![image](https://github.com/user-attachments/assets/66915189-5b6f-4e77-bd09-76e155dd3d7a)

![image](https://github.com/user-attachments/assets/c0abade1-58e8-480d-93cb-ad4ed60bcb72)

![image](https://github.com/user-attachments/assets/c43af8e0-2a8c-42a6-a6f8-b2e944bc9d04)
## Dev notes (Optional)
Try to follow the form pattern. The edit and the add button have the
same form. The goal is to be very user friendly
## Linked pull requests (Optional)